Before create name refactor
